utils/requirements.txt
packaging==23.1
Pillow==10.1.0
pyparsing==3.0.9
